Thrift types

+++ Dbarone.Net.Database/Core/Serialization/Parquet/Thrift/ColumnMetaData.cs
@@ -124,7 +124,7 @@ public sealed class ColumnMetaData
 /// enums are also used to specify the encoding of definition and repetition levels.
 /// See the accompanying doc for the details of the more complicated encodings.
 /// </summary>
-public enum Encoding : short
+public enum Encoding : byte
 {
   /// <summary>
   /// Default encoding.
